no longer using libLLVM when building a halo bin
libLLVM is great for keeping the executable sizes down.

main reasons for avoiding is that libLLVM.so is not produced in
the build dir, so it breaks out test suite. in addition, we had
an ugly way to set the rpath relative to the clang rt lib dir.

Instead, we just rely on the fact that clang in whichever bin/ dir
it lives in, knows where the lib/ dir is that contains the static
LLVM lib components, and always includes it as a lib path via -L
when compiling.