override the dtor of llvm::TaskQueue; overlay the queue.
We're overriding the dtor by reimplementing the TaskQueue because
it blocks waiting for all of the work in the ThreadPool to finish,
during which it is invalid to add more work to the ThreadPool.

I think this behavior is silly. One should be able to implement
a sequential task queue that utilizes a thread pool such that
upon destruction, it only ensures the tasks enqueued up to that
point have finished. The thread pool's other tasks are irrelevant.