[skip ci] working on dynlinker
skipping CI because we broke dependencies on static libs of LLVM
in this commit because halomon needs ORC JIT components now
that are missing.

Now is the time to fix the problems in issue #4 because
the current hack is extremely fragile and not easy to
update (llvm-config is not perfect).