finished the main fix for issue #7
The reason why we were not able to access globals
with dlsym linking was that the symbols in an executable
will only be available in the dynamic symbol table if
--export-dynamic ld flag is given.

This doesn't expose internal global values, like function
static globals. For those, we rely on the HaloPrepare pass
that is run by clang before assembly code is emitted.
Currently, this pass just exposes these function statics,
though the current implementation is prone to name clashes.